# app/api/loads/route.ts
// Loads API: Gets and creates cargo routes based on who is logged in.
import { NextRequest, NextResponse } from "next/server";
import dbConnect from "@/lib/mongodb";
import Load, { IStop } from "@/models/Load";
import { getUserFromRequest, requireRole } from "@/lib/auth";
import { logAction } from "@/lib/audit";

/**
 * ======================================================================================
 * API ROUTE: /api/loads
 * ======================================================================================
 * Handles the retrieval and creation of logistics loads.
 * Features:
 * 1. Role-Based Access: Drivers see only assigned loads; Dispatchers/Admins see all.
 * 2. Map Enrichment: Automatically geocodes addresses and calculates route stats (POST).
 * 3. Audit Logging: Tracks every load creation in the system audit trail.
 * 4. Concurrency: Uses MongoDB for persistence and real-time retrieval.
 * ======================================================================================
 */

export async function GET(req: NextRequest) {
  try {
    await dbConnect();
    const user = await getUserFromRequest(req);
    if (!user) {
      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    // ----------------------------------------------------------------------------------
    // FILTER LOGIC
    // ----------------------------------------------------------------------------------
    const filter: any = {};
    if (user.role === 'Driver') {
      // Drivers only see missions assigned directly to them
      filter.assignedDriverId = user.id;
    } else if (user.role === 'Dispatcher') {
      // Dispatchers see the global operational overview
    }
    // Admins see everything, including audit-ready cancelled loads

    const loads = await Load.find(filter)
      .populate('assignedDriverId', 'name email')
      .populate('createdBy', 'name email')
      .sort({ createdAt: -1 });
      
    return NextResponse.json(loads);
  } catch (error: unknown) {
    return NextResponse.json({ error: error instanceof Error ? error.message : "An error occurred" }, { status: 500 });
  }
}

export async function POST(req: NextRequest) {
  try {
    const user = await getUserFromRequest(req);
    if (!requireRole(user, ['Admin', 'Dispatcher'])) {
      return NextResponse.json({ error: "Forbidden: Only Dispatchers and Admins can create loads" }, { status: 403 });
    }

    await dbConnect();
    const body = await req.json();
    const { 
      loadNumber,
      commodity,
      pickups,
      deliveries,
      quantity,
      quantityUnit,
      weight,
      weightUnit
    } = body;

    // ----------------------------------------------------------------------------------
    // VALIDATION & MAP ENRICHMENT
    // ----------------------------------------------------------------------------------
    if (!loadNumber || !commodity || !pickups || !Array.isArray(pickups) || pickups.length === 0 || 
        !deliveries || !Array.isArray(deliveries) || deliveries.length === 0 || 
        !quantity || !quantityUnit || !weight || !weightUnit) {
      return NextResponse.json({ error: "Missing or invalid required fields (loadNumber, pickups, deliveries, quantity, weight)" }, { status: 400 });
    }

    // Check for duplicate load number
    const existingLoad = await Load.findOne({ loadNumber });
    if (existingLoad) {
      return NextResponse.json({ error: `Load Reference Number '${loadNumber}' is already in use.` }, { status: 400 });
    }

    const { getMockCoordinates, calculateMockRouteStats } = await import('@/lib/maps');

    const enrichedPickups = await Promise.all(pickups.map(async (p: IStop) => {
      const coords = await getMockCoordinates(p.address, p.city, p.state);
      return { ...p, date: new Date(p.date), lat: coords.lat, lng: coords.lng };
    }));
    
    const enrichedDeliveries = await Promise.all(deliveries.map(async (d: IStop) => {
      const coords = await getMockCoordinates(d.address, d.city, d.state);
      return { ...d, date: new Date(d.date), lat: coords.lat, lng: coords.lng };
    }));

    const routeStats = await calculateMockRouteStats(enrichedPickups, enrichedDeliveries);

    const newLoad = await Load.create({
      loadNumber,
      commodity,
      pickups: enrichedPickups,
      deliveries: enrichedDeliveries,
      totalDistance: routeStats.distance,
      estimatedDuration: routeStats.duration,
      quantity: Number(quantity),
      quantityUnit,
      weight: Number(weight),
      weightUnit,
      createdBy: user!.id,
      status: 'PENDING'
    });

    await logAction({
      req,
      userId: user!.id,
      action: 'LOAD_CREATED',
      entityType: 'Load',
      entityId: newLoad._id.toString(),
      details: { loadNumber },
    });

    return NextResponse.json(newLoad, { status: 201 });
  } catch (error: unknown) {
    return NextResponse.json({ error: error instanceof Error ? error.message : "An error occurred" }, { status: 500 });
  }
}