# openapi.yaml
openapi: 3.0.3
info:
  title: LoadFlow API
  version: 1.3.0
  description: |
    Welcome to the LoadFlow API!
    
    This API provides a complete backend for the LoadFlow Dispatch Application, allowing dispatchers, drivers, and administrators to orchestrate logistics seamlessly.

    ### 🔑 Authentication

    The system uses **Cookie-based JWT Authentication**.

    1. **Login** or **Register** (if invited) to receive a secure HTTP-only cookie.
    2. This cookie is automatically sent with subsequent requests.
    3. **Note for API Tools (Postman/Swagger):** Ensure cookies are enabled/stored after login so protected endpoints can be tested.

    ### 📦 Primary Features
    
    This API is built explicitly to handle fleet operations.
    
    - **Logistics Data:** Access resources in JSON format.
    - **Driver Dashboard Metrics:** Real-time calculation of Service Hours, Miles, and Completed Missions.
    - **Smart Notifications:** Optimistic state updates and ultra-fast polling for instant driver/dispatcher feedback.
    - **Proof of Delivery:** Secure `multipart/form-data` uploads handled and routed towards Cloudinary.
    - **Server Actions:** Public gateways directly dispatch SMTP NodeMailer notifications without persistence overhead.

    ### 👥 User Roles

    - **Driver:** Can view assigned configurations, update transit statuses, track performance metrics (Miles/Hours), and submit Proof of Delivery.
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/loads:
    get:
      tags: [Loads]
      summary: Get loads visible to the logged-in user
      description: |
        - Admin: all loads
        - Dispatcher: all non-cancelled loads
        - Driver: assigned non-cancelled loads only
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Load list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Load'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Loads]
      summary: Create a load
      description: Admin or Dispatcher only
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoadRequest'
      responses:
        '201':
          description: Load created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/loads/{id}/status:
    patch:
      tags: [Loads]
      summary: Update load status, stop status, or assignment fields
      description: |
        Supports three practical flows:
        - Driver stop updates using `stopType`, `stopIndex`, and `stopStatus`
        - Dispatcher/Admin whole-load status updates using `status`
        - Dispatcher/Admin assignment field updates using equipment and driver fields
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadStatusPatchRequest'
      responses:
        '200':
          description: Load updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
        '400':
          description: Invalid status transition or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Load not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/notifications:
    get:
      tags: [Notifications]
      summary: Fetch user notifications
      security:
        - cookieAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [ALL, LOADS, USERS]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: all
          in: query
          description: Include read notifications if true
          schema:
            type: boolean
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    patch:
      tags: [Notifications]
      summary: Mark all unread notifications as read
      security:
        - cookieAuth: []
      responses:
        '200':
          description: All marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: MongoDB document id
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
    Notification:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
          nullable: true
        targetRole:
          type: string
          enum: [Admin, Dispatcher, Driver]
          nullable: true
        message:
          type: string
        type:
          type: string
          enum: [INFO, WARNING, DANGER, SUCCESS]

        link:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Stop:
      type: object
      required:
        - address
        - city
        - state
        - postalCode
        - appointmentNumber
        - date
        - time
      properties:
        address:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        appointmentNumber:
          type: string
        date:
          type: string
          format: date-time
        time:
          type: string
        status:
          type: string
          enum: [PENDING, PICKED_UP, DELIVERED]
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
    CreateLoadRequest:
      type: object
      required:
        - loadNumber
        - commodity
        - pickups
        - deliveries
        - quantity
        - quantityUnit
        - weight
        - weightUnit
      properties:
        loadNumber:
          type: string
        commodity:
          type: string
        pickups:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Stop'
        deliveries:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Stop'
        quantity:
          type: number
        quantityUnit:
          type: string
          enum: [skids, pallets, packages, pieces, box, cases]
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        totalDistance:
          type: number
          nullable: true
          description: In miles (mock feature)
        estimatedDuration:
          type: number
          nullable: true
          description: In hours (mock feature)
...
    LoadStatusPatchRequest:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, ASSIGNED, IN_TRANSIT, PICKED_UP, DELIVERED, CANCELLED, COMPLETED]
        stopType:
          type: string
          enum: [pickups, deliveries]
        stopIndex:
          type: integer
        stopStatus:
          type: string
          enum: [PENDING, PICKED_UP, DELIVERED]
        assignedDriverId:
          type: string
          nullable: true
        truckNumber:
          type: string
          nullable: true
        trailerNumber:
          type: string
          nullable: true
        truckType:
          type: string
          enum: [Sleeper Cab, Day Cab]
          nullable: true
        trailerType:
          type: string
          enum: [Dry Van, Reefer, Tri Axle, Flatbed]
          nullable: true
        __v:
          type: integer
    Load:
      type: object
      properties:
        _id:
          type: string
        loadNumber:
          type: string
        commodity:
          type: string
        pickups:
          type: array
          items:
            $ref: '#/components/schemas/Stop'
        deliveries:
          type: array
          items:
            $ref: '#/components/schemas/Stop'
        quantity:
          type: number
        quantityUnit:
          type: string
          enum: [skids, pallets, packages, pieces, box, cases]
        weight:
          type: number
        weightUnit:
          type: string
          enum: [lbs, kg]
        truckType:
          type: string
          enum: [Sleeper Cab, Day Cab]
          nullable: true
        trailerType:
          type: string
          enum: [Dry Van, Reefer, Tri Axle, Flatbed]
          nullable: true
        trailerNumber:
          type: string
          nullable: true
        truckNumber:
          type: string
          nullable: true
        assignedDriverId:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
                email:
                  type: string
                  format: email
          nullable: true
        createdBy:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
                email:
                  type: string
                  format: email
        status:
          type: string
          enum: [PENDING, ASSIGNED, IN_TRANSIT, PICKED_UP, DELIVERED, CANCELLED, COMPLETED]
        podUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalDistance:
          type: number
          nullable: true
          description: In miles (mock feature)
        estimatedDuration:
          type: number
          nullable: true
          description: In hours (mock feature)

        __v:
          type: integer
...